modularizes weather
