+++ server.js
@@ -4,6 +4,7 @@ require('dotenv').config();
 const express = require('express');
 const cors = require('cors');
 const axios = require('axios');
+const handleGetWeather = require('./modules/weather'); //video said we don't need to add .js
 
 const app = express();
 
@@ -17,13 +18,6 @@ app.listen(PORT, () => console.log('server is listening on port', PORT));
 app.get('/weather', handleGetWeather);
 app.get('/movie', handleGetMovie);
 
-class Forcast {
-    constructor(obj){
-        this.description = `Low of ${obj.low_temp}, high of ${obj.max_temp} with ${obj.weather.description}`;
-        this.date = obj.datetime;
-    }
-}
-
 class Movie {
     constructor(obj){
         this.title = obj.title;
@@ -33,22 +27,6 @@ class Movie {
         this.image_url = `https://image.tmdb.org/t/p/w500/${obj.poster_path}`;
         this.popularity = obj.popularity;
         this.released_on = obj.release_date;
-    }
-}
-
-//forEach weather.data[{}] find .weather.description && .datetime
-
-async function handleGetWeather(request, response) {
-    const url = `http://api.weatherbit.io/v2.0/forecast/daily?lat=${request.query.lat}&lon=${request.query.lon}&key=${process.env.WEATHER_API_KEY}&units=I&days=12`;
-
-    try{
-        const result = await axios.get(url);
-        const weatherObjArr = result.data.data.map(dataObj => new Forcast(dataObj));
-        response.send(weatherObjArr);
-    } catch (error) {
-        response.status(204).send({
-            message: 'No weather data found for location!'
-        });
     }
 }
